Share the ForwardDiff tag between models and fix a typo in precompilation
Currently, every model gets its own ForwardDiff tag which means that every model
also have a unique type of their dual numbers. This causes every function called
with dual numbers to have to be recompiled for every model.

In this PR, we define a shared tag in ModelBasedEcon that all models use.
This means that we can push the precompile generation for many functions
from the model into ModelBaseEcon itself which changes the cost of them
from O(1) to O(n_models).

This PR also corrects a mismatch in the `precompile` call and the call to `ForwardDiff`.
In the precompile calls `MyTag` was used as the type to precompile for which means that the calls
to `GradientConfig` should have used `MyTag()` (so that the type of the tag was `MyTag`.) Now,
when `MyTag` was used to the `GradientConfig` call the type of it is actually `DataType` which
means that the types in the `precompile` call was different compared to the types actually
encountered at runtime.

Using the following benchmark script:

```julia
unique!(push!(LOAD_PATH, realpath("./models"))) # hide

@time using ModelBaseEcon
using Random # See https://github.com/JuliaLang/julia/pull/48810

@time using FRBUS_VAR

m = FRBUS_VAR.model
nrows = 1 + m.maxlag + m.maxlead
ncols = length(m.allvars)
pt = zeros(nrows, ncols);
@time @eval eval_RJ(pt, m);

using BenchmarkTools
@btime eval_RJ(pt, m);
```

this PR has the following changes:

- Loading ModelBaseEcon: 0.641551s -> 0.645943s
- Loading model 0.053s -> 0.032s
- First call `eval_RJ`: 5.50s -> 0.64s
- Benchmark `eval_RJ`:  597.966μs -> 573.923μs