make `LittleDict` types in structs concrete
The `LittleDict` type actually has 4 type parameters and all type parameters have to be specified in a field for it to be considered concrete.

The current version of the code only specifies two type parameters which means that type inference would be imprecise when dealing with these fields.

Also, this imprecise type inference made the code vulnerable to "invalidations". In practice, this had the effect while running `eval_RJ` would have low latency but if soe other package (in this case DataStructures) is loaded before this call it would be slow.

Here, if DataStructures was loaded, the first call to `eval_RJ` takes 3 seconds due to invalidations but with this PR it is back to ~0.07 seconds (timed on 1.9.0-beta4).